Refactor API View model
